requirements.txt
fastapi==0.100.0
python-dotenv~=1.0.0
requests~=2.31.0
uvicorn==0.22.0
email_validator==2.0.0